:art: Improving type inference
Previously, the Eq + Ord classes were designed such that type inference
wasn't as good as it could've been. Specifically, there was no inference
in a reverse direction: knowing that `x .== x :: Defined Bool` wouldn't
even be enough to tell me that `x` was wrapped in `Defined`. The
reformulation of these classes now allows that to happen, and hopefully
makes inference quite a lot better.

This closes #10.